+++ Editor/HeaderGUI/CacheAndUtil.cs
@@ -1,6 +1,5 @@
 // ReSharper disable once RedundantUsingDirective
 using System;
-// ReSharper disable once RedundantUsingDirective
 using UnityEditor;
 using System.Collections.Generic;
 using SaintsField.Editor.Core;